requirements.txt
aiogram==3.3.0
telethon==1.33.1
python-dotenv==1.0.1
rarfile==4.1 
pysocks==1.7.1
